fix: POS List Bank Accounts with mandatory Bank identifier.
