+++ src/api/ADempiere/form/VPOS/index.js
@@ -886,65 +886,6 @@ export function listStocks({
 }
 
 /**
- * List Banks
- */
-export function listBanks({
-  posId
-}) {
-  return request({
-    url: `point-of-sales/${posId}/banks`,
-    method: 'get',
-    params: {
-      page_size: 100
-    }
-  })
-}
-
-/**
- * List Bank Accounts
- */
-export function listBankAccounts({
-  posId,
-  customerId,
-  bankId
-}) {
-  return request({
-    url: `point-of-sales/${posId}/customers/${customerId}/bank-accounts`,
-    method: 'get',
-    params: {
-      page_size: 100,
-      bank_id: bankId
-    }
-  })
-}
-
-/**
- * Create Customer Account
- */
-export function createCustomerBankAccount({
-  posId,
-  customerId,
-  accountNo,
-  driverLicense,
-  bankId,
-  bankAccountType = 'C',
-  isAch = true
-}) {
-  return request({
-    url: `point-of-sales/${posId}/customers/${customerId}/bank-accounts`,
-    method: 'post',
-    data: {
-      account_no: accountNo,
-      driver_license: driverLicense,
-      bank_id: bankId,
-      bank_account_type: bankAccountType,
-      social_security_number: driverLicense,
-      is_ach: isAch
-    }
-  })
-}
-
-/**
  * Credit Memo as Payment Method
  */
 export function listCustomerCredits({
